+++ Samples/kotlin/src/main/java/com/phunware/kotlin/sample/location/LocationSharingActivity.kt
@@ -127,7 +127,7 @@ class LocationSharingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback,
             override fun onItemSelected(parent: AdapterView<*>, view: View, position: Int, id: Long) {
                 val floor = spinnerAdapter.getItem(id.toInt())
                 if (floor != null) {
-                    currentBuilding?.selectFloor(floor.level)
+                    currentBuilding?.selectFloor(floor.id)
                 }
             }
 
@@ -210,8 +210,7 @@ class LocationSharingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback,
                         // Animate the camera to the building at an appropriate zoom level
                         val cameraUpdate = CameraUpdateFactory
                                 .newLatLngBounds(initialFloor.bounds, 4)
-                        phunwareMap.googleMap.animateCamera(cameraUpdate)
-
+                        mapManager.animateCamera(cameraUpdate)
 
                         // Start sharing location with other users
                         startLocationSharing()