+++ Samples/kotlin/src/main/java/com/phunware/kotlin/sample/routing/CampusRoutingActivity.kt
@@ -219,16 +219,16 @@ open class CampusRoutingActivity : AppCompatActivity(), OnPhunwareMapReadyCallba
                                     floorSpinnerAdapter.addAll(sortedFloors)
                                     floorSpinnerAdapter.notifyDataSetChanged()
                                     var selectedFloorIndex = 0
-                                    val selectedFloor = currentBuilding.selectedFloor
-                                    if (selectedFloor != null) {
-                                        val selectedFloorOptions = getFloorOptionsFromSpinner(selectedFloor.id)
-                                        if (selectedFloorOptions != null) {
-                                            selectedFloorIndex = floorSpinnerAdapter.getPosition(selectedFloorOptions)
-                                            if (selectedFloorIndex == -1) {
-                                                selectedFloorIndex = 0
-                                            }
+                                    val selectedFloorOptions = selectedBuilding.selectedFloor?.id?.let {
+                                        getFloorOptionsFromSpinner(it)
+                                    }
+                                    if (selectedFloorOptions != null) {
+                                        selectedFloorIndex = floorSpinnerAdapter.getPosition(selectedFloorOptions)
+                                        if (selectedFloorIndex == -1) {
+                                            selectedFloorIndex = 0
                                         }
                                     }
+
                                     floorSpinner.setSelection(selectedFloorIndex)
                                 }
                             }
@@ -247,7 +247,7 @@ open class CampusRoutingActivity : AppCompatActivity(), OnPhunwareMapReadyCallba
                                         currentBuilding.selectFloor(currentFloorId)
                                     }
                                     val cameraUpdate = CameraUpdateFactory.newLatLngBounds(selectedFloor.bounds, 4)
-                                    phunwareMap.googleMap.animateCamera(cameraUpdate)
+                                    mapManager.animateCamera(cameraUpdate)
                                 }
                             }
 
@@ -255,13 +255,20 @@ open class CampusRoutingActivity : AppCompatActivity(), OnPhunwareMapReadyCallba
                         }
 
                         // Set building to initial floor value
-                        val initialFloor = currentBuilding.initialFloor
-                        currentBuilding.selectFloor(initialFloor.id)
+                        var buildingIndex = campus.buildings.indexOfFirst { building ->
+                            building.id == currentBuilding.id
+                        }
+
+                        if (buildingIndex == -1) {
+                            buildingIndex = 0
+                        }
+
+                        buildingSpinner.setSelection(buildingIndex, false)
 
                         // Animate the camera to the building at an appropriate zoom level
                         val cameraUpdate = CameraUpdateFactory
-                            .newLatLngBounds(initialFloor.bounds, 4)
-                        phunwareMap.googleMap.animateCamera(cameraUpdate)
+                            .newLatLngBounds(currentBuilding.initialFloor.bounds, 4)
+                        mapManager.animateCamera(cameraUpdate)
 
                         // Enabled fab for routing
                         showFab(true)